requirements.txt
aiohttp==3.7.4.post0
cchardet==2.1.7
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
httpie==2.4.0
psycopg2-binary==2.8.6
PyJWT==2.0.1
SQLAlchemy==1.4.6
python-dotenv==0.17.0
Flask-Migrate==2.7.0
Flask-Script==2.0.6
passlib==1.7.4
Flask-HTTPAuth==4.2.0
Werkzeug==1.0.1
